refactor: move volumes from ledger service to facade
