+++ src/services/ledger/facade/index.types.d.ts
@@ -1,3 +1,6 @@
+type TxnGroup = keyof typeof import("./volume").TxnGroups
+type TxnTypes = (typeof import("./volume").TxnGroups)[TxnGroup]
+
 type InitialStaticAccountIds = {
   bankOwnerAccountId: LedgerAccountId | Error
   dealerBtcAccountId: LedgerAccountId | Error
