+++ src/services/ledger/facade/volume.ts
@@ -1,6 +1,125 @@
+import {
+  LedgerTransactionType,
+  UnknownLedgerError,
+  toLiabilitiesWalletId,
+} from "@domain/ledger"
+
 import { timestampDaysAgo } from "@utils"
 
-import { TxVolumeAmountSinceFactory } from "../volume"
+import { paymentAmountFromNumber } from "@domain/shared"
+import { addAttributesToCurrentSpan } from "@services/tracing"
+
+import { Transaction } from "../books"
+
+export const TxnGroups = {
+  allPaymentVolumeSince: [
+    LedgerTransactionType.IntraLedger,
+    LedgerTransactionType.OnchainIntraLedger,
+    LedgerTransactionType.LnIntraLedger,
+    LedgerTransactionType.Payment,
+    LedgerTransactionType.OnchainPayment,
+  ],
+  externalPaymentVolumeSince: [
+    LedgerTransactionType.Payment,
+    LedgerTransactionType.OnchainPayment,
+  ],
+  intraledgerTxBaseVolumeSince: [
+    LedgerTransactionType.IntraLedger,
+    LedgerTransactionType.OnchainIntraLedger,
+    LedgerTransactionType.LnIntraLedger,
+  ],
+  tradeIntraAccountTxBaseVolumeSince: [
+    LedgerTransactionType.WalletIdTradeIntraAccount,
+    LedgerTransactionType.OnChainTradeIntraAccount,
+    LedgerTransactionType.LnTradeIntraAccount,
+  ],
+  lightningTxBaseVolumeSince: [
+    LedgerTransactionType.Payment,
+    LedgerTransactionType.Invoice,
+    LedgerTransactionType.LnFeeReimbursement,
+  ],
+  onChainTxBaseVolumeSince: [
+    LedgerTransactionType.OnchainPayment,
+    LedgerTransactionType.OnchainReceipt,
+  ],
+  allTxBaseVolumeSince: Object.values(LedgerTransactionType),
+} as const
+
+const TxVolumeAmountSinceFactory = () => {
+  const txVolumeSince = async ({
+    walletId,
+    timestamp,
+    txnGroup,
+  }: {
+    walletId: WalletId
+    timestamp: Date
+    txnGroup: TxnGroup
+  }): Promise<TxBaseVolume | LedgerServiceError> => {
+    const liabilitiesWalletId = toLiabilitiesWalletId(walletId)
+
+    const txnTypes: TxnTypes = TxnGroups[txnGroup]
+    const txnTypesObj = txnTypes.map((txnType) => ({
+      type: txnType,
+    }))
+
+    try {
+      const [result]: (TxBaseVolume & { _id: null })[] = await Transaction.aggregate([
+        {
+          $match: {
+            accounts: liabilitiesWalletId,
+            $or: txnTypesObj,
+            $and: [{ timestamp: { $gte: timestamp } }],
+          },
+        },
+        {
+          $group: {
+            _id: null,
+            outgoingBaseAmount: { $sum: "$debit" },
+            incomingBaseAmount: { $sum: "$credit" },
+          },
+        },
+      ])
+
+      const outgoingBaseAmount = result?.outgoingBaseAmount ?? (0 as CurrencyBaseAmount)
+      const incomingBaseAmount = result?.incomingBaseAmount ?? (0 as CurrencyBaseAmount)
+      addAttributesToCurrentSpan({
+        "txVolume.function": txnGroup,
+        "txVolume.outgoing": outgoingBaseAmount.toString(),
+        "txVolume.incoming": incomingBaseAmount.toString(),
+      })
+      return { outgoingBaseAmount, incomingBaseAmount }
+    } catch (err) {
+      return new UnknownLedgerError(err)
+    }
+  }
+
+  const create =
+    (txnGroup: TxnGroup): GetVolumeAmountSinceFn =>
+    async (args) => {
+      const {
+        walletDescriptor: { id: walletId, currency: walletCurrency },
+        timestamp,
+      } = args
+      const volume = await txVolumeSince({ walletId, timestamp, txnGroup })
+      if (volume instanceof Error) return volume
+
+      const outgoingBaseAmount = paymentAmountFromNumber({
+        amount: volume.outgoingBaseAmount,
+        currency: walletCurrency,
+      })
+      if (outgoingBaseAmount instanceof Error) return outgoingBaseAmount
+
+      const incomingBaseAmount = paymentAmountFromNumber({
+        amount: volume.incomingBaseAmount,
+        currency: walletCurrency,
+      })
+      if (incomingBaseAmount instanceof Error) return incomingBaseAmount
+
+      return { outgoingBaseAmount, incomingBaseAmount }
+    }
+
+  return { create }
+}
 
 const VolumesForAccountWalletsFactory = () => {
   const create =
@@ -35,16 +154,34 @@ const VolumesForAccountWalletsFactory = () => {
   return { create }
 }
 
-const volumesFactory = VolumesForAccountWalletsFactory()
+const volumeAmountForAccountFactory = VolumesForAccountWalletsFactory()
+const txVolumeAmountFactory = TxVolumeAmountSinceFactory()
 
-export const externalPaymentVolumeAmountForAccountSince = volumesFactory.create(
+export const externalPaymentVolumeAmountForAccountSince =
+  volumeAmountForAccountFactory.create("externalPaymentVolumeSince")
+export const externalPaymentVolumeAmountSince = txVolumeAmountFactory.create(
   "externalPaymentVolumeSince",
 )
 
-export const intraledgerTxBaseVolumeAmountForAccountSince = volumesFactory.create(
+export const intraledgerTxBaseVolumeAmountForAccountSince =
+  volumeAmountForAccountFactory.create("intraledgerTxBaseVolumeSince")
+export const intraledgerTxBaseVolumeAmountSince = txVolumeAmountFactory.create(
   "intraledgerTxBaseVolumeSince",
 )
-
-export const tradeIntraAccountTxBaseVolumeAmountForAccountSince = volumesFactory.create(
+export const tradeIntraAccountTxBaseVolumeAmountForAccountSince =
+  volumeAmountForAccountFactory.create("tradeIntraAccountTxBaseVolumeSince")
+export const tradeIntraAccountTxBaseVolumeAmountSince = txVolumeAmountFactory.create(
   "tradeIntraAccountTxBaseVolumeSince",
+)
+
+export const allPaymentVolumeAmountSince = txVolumeAmountFactory.create(
+  "allPaymentVolumeSince",
+)
+export const allTxBaseVolumeAmountSince =
+  txVolumeAmountFactory.create("allTxBaseVolumeSince")
+export const onChainTxBaseVolumeAmountSince = txVolumeAmountFactory.create(
+  "onChainTxBaseVolumeSince",
+)
+export const lightningTxBaseVolumeAmountSince = txVolumeAmountFactory.create(
+  "lightningTxBaseVolumeSince",
 )
